break up query file
